Restart theming process for the website, namely to change colors.
I have created a fork,
https://github.com/bbjornstad/zola-theme-termigawa, which is the
terminimal theme, but forked so as to use the Kanagawa.nvim color scheme
which is just too good. Also have worked heavily on page structure. The
given zola template terminimal is very nice, but perhaps a bit too
rigidly defined for my use case. A lot of the forks changes are probably
able to be asubsumed by the base website files herex.